Ubuntu Install: fix install script & update README
This patch implements a fix for the installation on the Ubuntu OS.
It removes the quotes from the NAME of the os-release file. These quotes
are not present in all distros, such as Fedora where they are absent.

Also, this patch fixes a typo in the README file concerning the py script
to call to run the dependancies installations.

Change-Id: I71c76a29ed1daee2b8ee133a20918d42a0156432
Tuleap: #451