initial documentation in RST format
This new documentation can be built using `make docs`. It should be
built automatically via a hook on github and published on
https://ring.readthedocs.org.

This allows for much better formatting and is more pleasant to browse
for new contributors. It is possible to build the docs using the
Makefile at the root of the repository.

Change-Id: I664d4662b76c525f2a67eac6266bada3e94aa835
Tuleap: #595