Packaging: Avoid conflicts with multiple concurrent builds
This patch helps to avoid conflicts with multiple concurrent builds.

Running package-debian8 two times on the same machine could have
resultet in the wrong docker image being used for a build, creating
confusing outputs.

We haven't hit that case yet, but it could happen if we run the weekly
packaging publishing job at the same time as we run tests.

Change-Id: I0272c510dec77f43867ff85a95bdac8fda038936
Tuleap: #872