deploy-packages: Don't fail when adding the same deb
When adding the same deb twice to the repository, deploy-packages fails.
This is due to reprepro complaining about adding two times the same deb
to the same architecture. This happens because we generate arch all
packages on both i386 and amd64. This patch tries to remove the package
before adding it, which won't do anything for most packages but will
solve the issue with arch all packages. This is much safer than ignoring
the error as there is no option to only ignore this type of error.

Change-Id: I397148f9e8efe6ca1a089644abbd86395fa5eaf4
Tuleap: #872