make-ring: bump qt5 version for OSX
We used qt 5.5.1 since 5.6 was causing errors in the OSX client.
Problem is the 5.5.1 compilation is broken on macOS 10.12 Sierra, This patch
separates the brew unlink process from the install/link process. It gives the
opportunity to be more expressive in the unlink process, in particular we can
unlink all the qt5 versions.

Tuleap: #1362
Change-Id: Iacfc5de47c2f799391a7cbbb40f38f86c5a4bc32