client-android: add ci cqfd
Add continuous integration tool cqfd. This tool uses a docker
environment to compile a project within a specific OS and given
specific dependencies.

It comes with a configuration file tailored for android, android_cqfdrc
It contains different flavors working for the android client (debug,
release, daemon-only)

cqfd relies on a dockerfile. This file will be in the client-android
repo.

Command example:
./cqfd -f .cqfd/android_cqfdrc -d client-android/ release

Change-Id: I3198389cc65246a2ce2636962da41980c79e2edf