build: return error value when executing script
make-ring is not returning the right exit status for any shell script
it executes.
By fixing its behaviour, this has highlighted an issue with the OSX
dependencies not being able to be updated with newer versions
This commit returns a fix value of 1 when a shell script returns an
error (exit status != 0) and make sure that every OSX dependency is
properly installed and updated
We will exit only if exitOnError is set to True in execute_script

We used qt 5.5.1 since 5.6 was causing errors in the OSX client.
Problem is the 5.5.1 compilation is broken on macOS 10.12 Sierra, This patch
separates the brew unlink process from the install/link process. It gives the
opportunity to be more expressive in the unlink process, in particular we can
unlink all the qt5 versions.

Tuleap: #1362
Change-Id: Iced6d2eeb0d3028a3935e433c461fd20375e2d0b