gentoo build update
Gentoo is a rolling-release distribution, so from time to time core packages will be updated (in this case glibc), which will change some
build behaviour. So in this regard it might be more fragile than other distributions with a different release pattern.

To avoid this happening too often I also added an _emerge --sync_ before the build of a package. This way the repository should always be
up-to-date and the base container doesn't have to be rebuilt too often.

* glibc has a sandbox violation while building in a docker container (working around it by disabling the sandbox for this package)
* layman has changed its behaviour and ask interactively for non official repoitories (changing config file via sed)
* in make.conf dummy video card is an optimization to not build all graphics drivers for a headless build container
* local.gen is a build optimization to not build all languages for glibc
* in the  portage/ folder the files are reflecting the current state of dependencies needed for ring to build

These changes are all necessary to make automatically a Gentoo package in a docker container and ultimately to run this in the CI
infrastructure

This can be used with _make -f Makefile.packaging package-gentoo_ like any other package build. At the moment it might depend on other
reviews being included. The first time the build can take a long time as the container is compiled and all dependencies are compiled.

Change-Id: I823ce1f1e67a40519a224977a8f8a4a6141a6e05