add Gentoo support to make-ring.py
Before the modification the script blocked its execution on Gentoo,
so you couldn't even update the subfolders via the --init option.

This is now possible.

There is also support for the --dependencies option. It will install
the current known dependencies from the ring-overlay for the daemon.
It assumes you have the overlay setup on your machine.

Change-Id: I82ddf30bdbb7c5b560fca7fc85b45434e4d17021