add interactive switch for docker run
When I run the *-interactive make target it opens a bash but it
blocks after the first command has been entered. According to the
docker help the -i switch stands for interactive and is necessary
for interactive executions.
With this enabled I no longer have the problem described above.

The interactive mode is only used for debugging while setting up a
container, so I think it's not used very much.

I have an easy workaround, but I think it could be handy for the next
guy who wants to add an automatic build.

Change-Id: Ia4b497175a11a566755ef82532f3475aebe79ff0