win32: fix build on Fedora >26
Default QMake configuration in mingw packages has changed since
Fedora 26 and some modifications are required in order to make the
ring client able to compile on it:
 - Change QMake spec from "win32-g++" to "mingw-w64-g++" which links
   with libqt5main.dll instead of the wrong libqtmain.dll.

With this patch, the client shouldn't be able to compile anymore on
Fedora <26. The corresponding patch in "client-windows" is also
required as parts of the fix include modifications of the project
file.

Change-Id: I7c8c6ac028c1bccc824691e8974a93569fb6bb9c
Reviewed-by: Olivier Soldano <olivier.soldano@savoirfairelinux.com>