packaging: install repo and keys on user agreement
Currently it is very difficult for non-power-users to get Ring
automatically updated when using the Debian package from the website.

This is because configuring the Ring package repository requires
command line operations.

We want the Ring Debian/Ubuntu/whatever-Debian-based package (the
one from the website, not the official one from the Debian archive)
to ask the user whether he agrees to register Ring's repo and GnuPG
key as trusted peers. On user agreement (and only in this case, we
don't want to be invasive) we set up everything automatically so that
Ring will get updated along with system updates.

In order to do that, we have to change some things from the official
package we have in the Debian archive:

* we need to provide both daemon and client in a single package.
Otherwise we will not be able to provide a one-click install (in fact,
users would have to install daemon and client in two separate steps
and that doesn't make much sense to non-geek users).

* we need to add prerm and postinst scripts which will take care of
installing the repo and keys (and this has to be done in different
ways depending on the distribution)

Also, remove transition package ring-gnome which has already
disappeared from the Debian package some time ago.

Change-Id: I34d5e5ba000d5d750c8db55c65b536a237a738b5