packaging: install PPA and keys on user agreement
Currently it is very difficult for non-power-users to get automatic
updates with Ring when using Ring's PPA.

This is because configuring the PPA for automatic updates requires
(simple, but, well...) command line operations.

We want the Ring Debian/Ubuntu/whatever-Debian-based package (the
one from the website, not the official one from the Debian archive,
uh) to ask the user whether he agrees to put Ring's PPA and GnuPG
key to the trusted peers. On user agreement (and only in this case,
we don't want to be invasive) we set up everything automatically so
that Ring will get updated from the PPA when the user requests system
updates.

Change-Id: I34d5e5ba000d5d750c8db55c65b536a237a738b5