packaging: build again ring client binary package
Recent packaging changes removed the ring binary package from ring.cx
and replaced it by ring-all. This is because it doesn't make sense to
most end users to install daemon and client separately. However this
change has broken compatibility with the previous packaging policy, at
the very least because we didn't provide a transition package.

There are two solutions:
1. Provide a transition package
2. Build ring binary package again

Both should work well but 2. has better compatibility with official
Debian/Ubuntu packaging so we opt for this one.

Change-Id: I556ae580126038de7077f91b137adbf1046d59e4