packaging: fix issues with oci on Linux Mint
ring-all automatically registers

dl.ring.cx/ring-nightly/{id}_{version_id}

(id and version_id from os-release) as a trusted source. This is
perfectly fine as long as the package is installed on an officially
supported system, but if the package is installed on an unsupported
system (e.g. Linux Mint) the installed source will be wrong even if
even if the package is expected to work.

This patch addresses this issue by adding if else if else statements
handling all possible cases. This is not as pretty as before but it
makes sure we handle all cases properly.

Change-Id: I25167e013c56b212f7d59275419921cd0732de17
Gitlab: #490