packaging: different releases for website
It has been decided that we would have different releases between
website and apt repository. The version from the website would have
debian revision -0 and the one from apt repositories -1.

The main difference resides in the fact that the version from the
website distributes a postinst script which installs key and apt
source.list so that Ring gets automatically updated when installed
via one click install. This is not the case with the one from apt
repos.

Change-Id: I7bbfdfa41b0133ebe93f16bc550aa437c8ddfa6a