packaging: renaming packages from ring to jami
This commit rename the tarball, debian, ubuntu and fedora packages.

This commit also modify the debian post install script to add support
our repositories for Debian 10 and Ubuntu 18.10 and 19.04 and create
transition packages to ensure a proper nama transition in debian based
distributions.

Warning: when using reprepo to add the ring transition deb package
this get into pool/main/j/jami. for now we think is a reprepo error
because reads the Source: instead of Package: field.

Change-Id: I47449a79c1a8bfd6ee0f3ffed618886f1dc1645a