packaging: renaming packages from ring to jami
This commit rename the tarball, debian, ubuntu and fedora packages.

This commit also modify the debian post install script to add support
our repositories for Debian 10 and Ubuntu 18.10 and 19.04.

packaging: transition package for debian based

Warning: when using reprepo to add the ring deb package in a repository
this get into pool/main/j/jami. for now we think is a reprepo error
because reads the Source: instead of Package: field.

Change-Id: I421fa6ea66f5fd060e66ea5a0725efa24e79ebd4