packaging: adds debian 10 and raspbian arm
Adds:
* Debian 9 armhf
* Debian 9 arm64
* Debian 10 armhf
* Debian 10 arm64
* Raspbian 10 armhf
* The One Click Install (oci) version for all this items

This commit also fix the --generate option in the generated makefile by
scripts/make-packaging-target.py and remove the --architecture option
there because was not used at any part of the script.

Because our builder docker can not docker build with ARM we implemented
a workaround called dirty qemu-static is documented and we have a issue
to sovle it.

The pipeline for packaging-deploy-gnulinux-generic got changed to add
the support for the ARM architecture but the OCI build is not enabled
because the build including the oci takes long (5 hours or plus).

Warning: require companion changes in daemon (topic debian-arm)
to ensure the right compilation flags

Change-Id: I905bbd4660824670d7de836129886f3d7f1dc83b