packaging: debian: use cross compilation for armhf platforms
Using qemu in user mode to build jami package for raspbian platforms
adds a lot of overhead.

Remove the virtualization using the Debian 10 cross building tools. The
host platform (i.e. the platform on which the package will be run) has
to be set and some compilation options has to be added in the
debian/rules file.

Also add the armhf architecture in the Docker container in order to
install the package dependencies for this platform.

Change-Id: Ic2282ad9377d52482bca4a4f0ef017fbcc1f8eb7