build: Add support to use cached tarballs for packaging tagets.
Previously, the only way to ensure a clean release tarball was to
start from scratch, fetching them all from the network.  This change
leverages the new 'list-tarballs' targets of the contrib build system
to copy only the relevant ones from the cache, if a cache is used.

It also constructs the pristine source release archive via
git-archive, which is more robust against including unwanted files,
and does away with the need to copy the cache source tarballs.

Change-Id: I4993b269f3b97f6a4b6b8592aa8b5f4eb5448943