build: Add support to use cached tarballs for packaging targets.
Previously, the only way to ensure a clean release tarball was to
start from scratch, fetching all the contrib sources from the
network (> 1 GiB).  This change leverages the new 'list-tarballs'
targets of the contrib build system to allow reusing the relevant
tarballs from the cache when available, which translates to a build
time about 15% faster.

It also constructs the pristine source release archive via
git-archive, which is more robust against including unwanted files.
Since only the required tarballs are included in the source archive,
its size is reduced from 1.4 GiB to 72 MiB.

Change-Id: I4993b269f3b97f6a4b6b8592aa8b5f4eb5448943