Update patch set 13
Patch Set 13:

> Patch Set 13: Code-Review-2
> 
> Can be removed from the stack as make-ring.py is there for init and checkout master.
> 
> If it works, https://stackoverflow.com/questions/1777854/how-can-i-specify-a-branch-tag-when-adding-a-git-submodule/18799234#18799234 can be used as submodule set branch was added in Q3 2019

If we always update to the latest code for each submodule, we loose the information useful for reproducibility. As for git submodules, their 'version' is always recorded as a commit hash; the following branch feature is just so that when you run 'git submodule update' it'll take the commit from that branch.  That's already what happenss when doing 'git submodule update --remote'.

The submodules fetched by default are severily outdated; there's no harm to bump them here IMO, especially since it corresponds to a validated state of the repo (w.r.t. packaging).

Patch-set: 13