build.py: Allow skipping the operating system check.
While it's not possible to automatically check and configure the
dependencies for unknown/unsupported distributions, build.py
still provides a convenient experience to checkout the sources and
install Jami locally.  In this mode of operation, the user is expected
to configure their environment by their own means.

* build.py: Rename Ring to Jami in the document string.  Remove
unused 'signal' import.
(choose_distribution): Do not crash when the '/etc/os-release' does
not exist; return 'Unknown' instead.
(validate_args): Allow disabling the distribution check when the user
provides 'no-check' as the value of the --distribution argument, or
defines the JAMI_BUILD_NO_CHECK environment variable.  Update the
help documentation.
* scripts/install.sh (priv_install): Document the arguments.

Change-Id: I41ba2da05771feb6e3f9bc7c087206596063ba81