Jenkinsfile: Move the Guix environment specification to a manifest file.
This makes the Jenkinsfile cleaner and allows developers to easily
reproduce the same environment.

* guix/minimal-manifest.scm: New file.
* Jenkinsfile ('Generate release tarball): Use it to spawn the
container used to build the release tarball.  Also expose /usr/bin/env
in the container.  Use the new 'portable-release-tarball' target to
simplify things.  Fix the shebang line; it needs to be the first line
of the script.

GitLab: https://git.jami.net/savoirfairelinux/jami-packaging/-/issues/98
Change-Id: I2b66cb0d54c3aa5f1f95ed2b499a370c935e0db6