build: Improvements to the generation of the release tarballs.
The phony tarball-producing targets will now always clear the
tarballs.manifest file before generating a new release tarball,
ensuring it doesn't contain stale information.

A new 'portable-release-tarball' target is also added to offer an easy
to use target that was previously implemented ad-hoc in the CI.  Move
this logic here as a portable release tarball is useful for packagers
wanting to test locally with the various supported platforms.  The
Jenkinsfile now makes use of it.

* Makefile (daemon/contrib/native/Makefile): New target.
(tarballs.manifest): Add the above target as a prerequisite and remove
invocation to bootstrap script.  Invoke make with --no-print-directory
to prevent entering/leaving directory messages from appearing in the
output (it's supposed to be on when using --silent, but when the make
invocation is nested, it doesn't seem to be the case).
(release-tarball): Also clear the tarballs.manifest when invoked.
(purge-release-tarballs, portable-release-tarball): New targets.
* guix/minimal-manifest.scm: New file.
* Jenkinsfile ('Generate release tarball): Use it to spawn the
container used to build the release tarball.  Also expose /usr/bin/env
in the container.  Use the new 'portable-release-tarball' target to
simplify things.  Fix the shebang line; it needs to be the first line
of the script.

GitLab: https://git.jami.net/savoirfairelinux/jami-packaging/-/issues/98
Change-Id: I3b55e25933108ae9930bc9fcc867f9fa51d796f2