packaging/fedora: Split into multiple packages.
Tackle the growing complexity of the packaging by splitting the
problem into smaller parts.  This change makes the daemon its own
package, along the client library and the GNOME client.

The build script is updated to reflect these changes, and
unused/obsolete spec files are removed from the project.

GitLab: #264
Change-Id: I5813e48e0c1248875a7a8f0ee1ec6fb2b87faef4