packaging: Add jami-libqt to the jami-all package.
This makes the current 'jami-all' package the one used for the
one-click install package.

* packaging/rules/debian/rules (OCI_INSTALL_DIR): New variable.
(override_dh_auto_install): Extend rule to install the files of the
jami-libqt package.  Do not include the jami-gnome in jami-all.
* packaging/rules/debian/control (jami-all): Update description.
* packaging/rules/debian-one-click-install: Delete directory.
* packaging/rules/debian-one-click-install/jami-all.postinst: Move to ...
* packaging/rules/debian/jami-all.postinst: ... here.
* scripts/make-packaging-target.py: Remove all OCI-specific targets.
* scripts/deploy-packages.sh (package_deb) [jami-all]: Do not add
package to the apt repository.
[manual-download]: Look for the jami-all package from the main
distribution directory, not in a '_oci' suffixed version.

GitLab: https://git.jami.net/savoirfairelinux/jami-client-qt/-/issues/263
Change-Id: Ie3c6bb024cfcf98f7ceffabcd6ecd17ffdfa41d9
* scripts/deploy-packages.sh (package_deb):