build: Improvements to the generation of the release tarballs.
The phony tarball-producing targets will now always clear the
tarballs.manifest file before generating a new release tarball,
ensuring it doesn't contain stale information.

A new 'portable-release-tarball' target is also added to offer an easy
to use target that was previously implemented ad-hoc in the CI.  Move
this logic here as a portable release tarball is useful for packagers
wanting to test locally with the various supported platforms.

* Makefile (daemon/contrib/native/Makefile): New target.
(tarballs.manifest): Add the above target as a prerequisite and remove
invocation to bootstrap script.  Invoke make with --no-print-directory
to prevent entering/leaving directory messages from appearing in the
output (it's supposed to be on when using --silent, but when the make
invocation is nested, it doesn't seem to be the case).
(release-tarball): Also clear the tarballs.manifest when invoked.
(purge-release-tarballs, portable-release-tarball): New targets.

Change-Id: I3b55e25933108ae9930bc9fcc867f9fa51d796f2