build: Keep tarballs.manifest in sync with contrib.
Previously, someone would need to remember to delete the
tarballs.manifest to have it regenerated, which was error prone.  This
change causes the 'tarballs.manifest' file to be recomputed when the
list of used contrib tarballs is modified in any way.

* Makefile (daemon/contrib/native/Makefile): New target.
(tarballs.manifest): Add the above target as a prerequisite and remove
invocation to bootstrap script.  Conditionally rebuild the target only
when its content has changed.

Change-Id: Icc3d07eae06dd9f3209fd0120bc1f6c3f2ec6272