packaging: debian: Do not alter the source tarball.
AFAICT, this was made to clear some supposedly non-redistributable
files in the sources according to Debian policy.  If there's some
cleanup to do, we should do so at the level of the common source
tarball, not in a Debian-specific way.

* packaging/rules/debian/rules (override_dh_auto_configure): Remove
code repacking tarballs.
(override_dh_clean, get-orig-source): Remove.
* packaging/rules/debian/copyright (Files-Excluded): Remove.  Also
streamline the listing of copyrights to the files part of our tree,
not in bundled libraries.  Debian-specific requirements can stay in
the Debian-maintained corresponding packages.
* scripts/build-package-debian.sh: Streamline, making the release
tarball the sole input of the process.  Do not build source packages,
for simplicity (dpkg-source doesn't like our source tarball as-is).

Change-Id: I78a482ab3362e9bba8ffdc0ab60324f94715bfb8