packaging: debian: Do not alter the source tarball.
AFAICT, this was made to clear some supposedly non-redistributable
files in the sources according to Debian policy.  If there's some
cleanup to do, we should do so for all the packages, not in a
Debian-specific way.

* packaging/rules/debian/rules (override_dh_auto_configure): Remove
code repacking tarballs.
(override_dh_clean): Remove.
(get-orig-source): Do not alter the source tarball.

Change-Id: I78a482ab3362e9bba8ffdc0ab60324f94715bfb8