packaging: Remove unused targets.
This further reduces the targets set to what the CI currently uses.
The native ARM (armhf, aarch64) builds are dropped as we are not
equipped in terms of native hardware to build them and QEMU user
emulation is too slow.  The remaining i386 builds are removed as they
are not used by the CI and have or are being phased out by their
upstreams.

* scripts/make-packaging-target.py
(DPKG_BASED_SYSTEMS_DOCKER_RUN_OPTIONS): New variable.
[debian_10]: Use it.
[debian_10_i386, debian_10_armhf, debian_10_arm64]
[ubuntu_18.04_i386, ubuntu_18.04_qt_i386]: Delete targets.
[debian_testing]: Use the above variable.
[ubuntu_20.04, ubuntu_20.10, ubuntu_21.04]: Likewise.
* docker/Dockerfile_debian_10_arm64: Delete file.
* docker/Dockerfile_debian_10_armhf: Likewise.
* docker/Dockerfile_debian_10_i386: Likewise.
* docker/Dockerfile_ubuntu_18.04_i386: Likewise.
* docker/Dockerfile_ubuntu_18.04_qt_i386: Likewise.

Change-Id: Iabb840e7f9a2111afe822e2c05797a651d55e21d