Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > Note that this means none of the contribs will have debug symbols available.  I think it should be commented out with a TODO to fix the assertions made by pjproject.  pjproject is *not* built with NDEBUG by default, which means on some distributions that provide their own pjproject (it used to be the case for Guix), these assertions may bite.
> 
> if you need debug on a particular lib, just recompile it with the options you need. this adds debug flags for all libs even the one who needs to be performent like ffmpeg and can cause a lot of unwanted behavior.

I'd argue that when you choose --debug that's a side effect you should be expecting anyway (lesser performance), and that the more important point is to be able to have informative backtraces and the ability to step things.  Perhaps we could have '--debug=all' or '--debug=main', with --debug=main being the default and --debug=all also turning debug on for the contribs.

Patch-set: 2