WIP: docker: Allow interrupting packaging jobs with SIGINT.
Use the recommended exec syntax for the CMD directive in the
Dockerfiles, which invokes the command directly rather than through a
shell.  Also use the --init option, which uses tini as minimal PID 1
init to proxy the signals to the launched processes; this helps as PID
1 is special and the kernel refuses to kill it.

* scripts/make-packaging-target.py: Provide the --init argument to
'docker run'.

The rest of this change was automated using the command:

  $ git grep -l CMD | xargs sed -i -E 's,CMD (.*)$,CMD ["\1"],g'

Change-Id: I93ba9ad8115541a7c7f02421384ae01a5b0650ad