packaging: Add the deb-pack-x86_64 and deb-pack-i686 targets.
The deb-pack targets build a Guix pack of the jami-qt client in the
.deb format, ready to be installed on any dpkg-based GNU/Linux
distribution.

* Makefile (RELEASE_VERSION, RELEASE_TARBALL_FILENAME): Export
variables.
(has-guix-p): New target, to be used as a predicate.
(portable-release-tarball): Rewrite in terms of the above.
(SUPPORTED_GNU_ARCHS, DEB_PACKS, DEB_PACK_TARGETS): New variables.
(guix-pack-command, define-deb-pack-rule): Likewise.
(deb-pack-x86_64, deb-pack-i686): New targets.
(PACKAGE-TARGETS): Register them.
* guix/guix-pack-manifest.scm: New file.
* guix/guix-pack-deb.postinst: Likewise.
* guix/guix-pack-deb.triggers: Likewise.
* scripts/deploy-packages.sh (package): Simplify and register the
'guix-deb-pack' distribution.
* Jenkinsfile (Sign & deploy packages): Simplify to deploy all the
packages subdirectories.  This makes it easier to handle multi-arch
packages for the same distribution, and removes the need to encode the
distribution information in the target name.

Change-Id: Id6f69e7a5872a39b4b8f50fa09c8ec3475e9c896