Makefile: Repatriate the definition of packaging targets.
Prior to this change, the Makefile packaging targets were generated by
a Python script, supposedly to reduce boilerplate.  Make is able to
programmatically define rules, so use its features instead.

The packaging targets 'package-' prefix is dropped, making the targets
names match the distribution directly, such as 'ubuntu-20.04'.  Only
hyphens are used in the package target names, making it easier to
remember.

The docker *-ro bind mounts are dropped; instead only the release
tarball file name is exposed in the containers as *the* source, which
makes it unambiguous.

The files related to RHEL builds are removed, as the RHEL build is
known to be broken and isn't being used.

* Makefile (DISTRIBUTIONS): New variable.
(make-docker-package-target): New function.
(docker/Dockerfile-snap): Adjust accordingly.
docker/*: Standardize to use hyphens instead of a mix of hyphens and
underscores.
* docker/Dockerfile_rhel_8: Delete file.
* docker/profile-seccomp-fedora_28.json: Delete file.
* scripts/make-packaging-target.py: Delete file.
* Jenkinsfile: Adjust accordingly.

Change-Id: I2dc42b489a54b177ad038398a9d23a2a7e3007ec