Makefile: Revert to use the last commit date.
The problem worked around
in (0e576761f44a82230d4cf2a9cd59585ee37ce862) by injecting the current
date earlier was with the CI not committing the adjusted submodules
before generating the tarball.  Using the date of the commit is better
as it doesn't change spuriously and is based on something more stable
that represents the state of the sources.

Change-Id: I03bc5513286e7aad891cc2ba94d2a1bb49e6eeb2