build: add separate lrc-gnome submodule for client-gnome
After the recent merge of the video renderer changes into lrc and
client-qt, the build for client-gnome is currently broken because it
has not yet been updated to follow the new API, a non-trivial task.
Thus, we add a separate lrc-gnome submodule for use by client-gnome,
as a temporary measure until the future vendoring of lrc into the two
clients.

With this change, the users of the Qt client ('jami' package) would
continue using 'jami-libclient' as a dependency, whereas for users of
the GNOME client ('jami-gnome') they would install and use the new
'jami-libclient-gnome' package.

Change-Id: I6aacb6c495c9a2953c328f2e9bfe6acd2b8cc645