packaging: fix libclient dependency issues of deb packages
Since for our deb packages all the binary packages share the same
source package, and are built in the same environment, depending on
the distro and dpkg tools versions, dpkg-shlibdeps might incorrectly
add the wrong libclient package for either clients.  For example,
have the jami-gnome package depend on both jami-libclient-gnome and
jami-libclient; which is wrong, since jami-libclient-gnome and
jami-libclient conflict with each other and cannot be installed
simultaneously.  So, we instruct dpkg-shlibdeps to exclude the two
libclient packages from ${shlibs:Depends}, which is fine because
we already manually add the correct one to each client's Depends.

Also, use the system libarchive-dev package for Ubuntu 18.04 as well;
since like with Debian 10, building libarchive from daemon's contribs
fails there too.

GitLab: #1415
Change-Id: I1c065a7fd3fc58324c7893d5d23039f2a5109931